refactor: move contributor languages into separate hook, clean up repoList usage
