+++ components/organisms/ContributorProfilePage/contributor-profile-page.tsx
@@ -3,7 +3,7 @@ import Text from "components/atoms/Typography/text";
 import ContributorProfileHeader from "components/molecules/ContributorProfileHeader/contributor-profile-header";
 import { ContributorObject } from "../ContributorCard/contributor-card";
 import CardLineChart from "components/molecules/CardLineChart/card-line-chart";
-import CardRepoList, { RepoList } from "components/molecules/CardRepoList/card-repo-list";
+import CardRepoList from "components/molecules/CardRepoList/card-repo-list";
 import PullRequestTable from "components/molecules/PullRequestTable/pull-request-table";
 import SkeletonWrapper from "components/atoms/SkeletonLoader/skeleton-wrapper";
 
@@ -17,7 +17,6 @@ import ContributorProfileTab from "../ContributorProfileTab/contributor-profile-
 import ProfileLanguageChart from "components/molecules/ProfileLanguageChart/profile-language-chart";
 import useFollowUser from "lib/hooks/useFollowUser";
 import useSupabaseAuth from "lib/hooks/useSupabaseAuth";
-import { getAvatarByUsername } from "lib/utils/github";
 
 const colorKeys = Object.keys(color);
 interface PrObjectType {
@@ -75,15 +74,8 @@ const ContributorProfilePage = ({
   });
 
   const { user: loggedInUser, signIn } = useSupabaseAuth();
-  const { chart, data } = useContributorPullRequestsChart(githubName, "*", repositories);
-  const repoList: RepoList[] = Array.from(new Set(data.map((prData) => prData.full_name))).map((repo) => {
-    const [repoOwner, repoName] = repo.split("/");
+  const { chart, repoList } = useContributorPullRequestsChart(githubName, "*", repositories);
 
-    return {
-      repoName: repoName,
-      repoIcon: getAvatarByUsername(repoOwner),
-    };
-  });
   const prsMergedPercentage = getPercent(prTotal, prMerged || 0);
   const { data: Follower, isError: followError, follow, unFollow } = useFollowUser(user?.login || "");
